requirements.txt
apache-flink==1.12.0
kafka-python==2.0.2
matplotlib==3.3.3
numpy==1.19.3
